# api/swagger.json
{
    "openapi":"3.0.2",
    "info": {
      "title":"API Title",
      "version":"1.0"
    },
    "servers": [
      {"url":"http://localhost:3000"}
    ],
    "components": {
      "schemas": {
        "cinema": {
          "type": "object",
          "required": [
            "id", "Name", "Address", "ProvinceCity"
          ],
          "properties": {
            "id": {
              "type": "integer",
              "description": "The auto-generated id of cinema"
            },
            "name": {
              "type": "string",
              "description": "Name of cinema"
            },
            "address": {
              "type": "string",
              "description": "Address of cinema"
            },
            "provinceCity": {
              "$ref": "#/components/schemas/provinceCity",
              "description": ""
            }
          },
          "example": {
            "id": 1, 
            "name": "CGV VinCom",
            "address": "123 lalalal",
            "provinceCity": {
              "id": 1,
              "name": "Da Nang"
            }
          }
        },
        "provinceCity": {
          "type": "object",
          "required": [
            "id", "Name"
          ],
          "properties": {
            "id": {
              "type": "integer",
              "description": "The auto-generated id of province/city"
            },
            "name": {
              "type": "string",
              "description": "Name of province/city"
            }
          }
        },
        "screen": {
          "type": "object",
          "required": [
            "id", "seatMatrix", "cinemaId"
          ],
          "properties": {
            "id": {
              "type": "integer",
              "description": "The auto-generated id of screen"
            },
            "seatMatrix": {
              "type": "string",
              "description": "Path of file seat matrix"
            },
            "cinemaId": {
              "type": "integer",
              "description": "The id of cinema"
            }
          },
          "example": {
            "id": 1,
            "seatMatrix": "./SeatMatrix/screen1.txt",
            "cinemaId": 1
          }
        },
        "cinemaDto": {
          "type": "object",
          "required": [
            "id", "name", "address", "provinceCityId"
          ],
          "properties": {
            "id": {
              "type": "integer",
              "description": "The auto-generated id of cinema"
            },
            "name": {
              "type": "string",
              "description": "Name of cinema"
            },
            "address": {
              "type": "string",
              "description": "Address of cinema"
            },
            "provinceCityId": {
              "type": "integer",
              "description": "Id of province/city"
            }
          },
          "example": {
            "id": 1, 
            "name": "CGV VinCom",
            "address": "123 lalalal",
            "provinceCityId": 1
          }
        },
        "show": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "description": "The auto-generated id of show"
            },
            "filmId": {
              "type": "integer",
              "description": "The film's id of this show"
            },
            "screenId": {
              "type": "integer",
              "description": "The screen's id of this show"
            },
            "timeStart": {
              "type": "string",
              "description": "The start time of this show"
            },
            "price": {
              "type": "integer",
              "description": "The price of this show"
            },
            "seatMatrix": {
              "type": "string",
              "description": "Path of file seat matrix"
            },
            "formatId": {
              "type": "integer",
              "description": "The id of format"
            },
            "subtitle": {
              "type": "boolean",
              "description": "True is have subtitle, false is have not subtile"
            },
            "dubbing": {
              "type": "boolean",
              "description": "True is have dubbing, false is have not dubbing"
            }
          }
        }
      },
      "parameters": {
        "CinemaPath": {
          "name": "cinemaId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer"
          },
          "description": "Cinema's id"
        },
        "ScreenPath": {
          "name": "screenId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer"
          },
          "description": "Screen's id"
        }
      },
      "requestBodies": {
        "cinemaBody": {
          "description": "Cinema information",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/cinemaDto"
              }
            }
          }
        },
        "screenBody": {
          "description": "Screen information",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/screen"
              }
            }
          }
        }
      }
    },
    "paths": {
      "/cinema": {
        "get": {
          "tags": [
            "Cinema"
          ],
          "description": "Get list cinemas from database",
          "parameters": [
            {
              "name": "provinceId",
              "in": "query",
              "schema": {
                "type": "integer"
              },
              "required": false,
              "description": "Province/city id to get list cinema by province/city"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/cinema"
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Resquest"
            },
            "500": {
              "description": "Server Error"
            }
          }
        },
        "post": {
          "description": "Add cinema into database",
          "tags": ["Cinema"],
          "parameters": [],
          "requestBody": {
            "$ref": "#/components/requestBodies/cinemaBody"
          },
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/cinema/{cinemaId}": {
        "get": {
          "tags": [
            "Cinema"
          ],
          "description": "",
          "parameters": [
            {
              "$ref": "#/components/parameters/CinemaPath"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "aplication/json": {
                  "schema": {
                    "$ref": "#/components/schemas/cinema"
                  }
                }
              }
            },
            "404": {
              "description": "Not Found"
            },
            "500": {
              "description": "Server Error"
            }
          }
        },
        "delete": {
          "description": "Delete cinema from database",
          "tags": ["Cinema"],
          "parameters": [
            {
              "$ref": "#/components/parameters/CinemaPath"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "put": {
          "description": "Update cinema informaiton ",
          "tags": ["Cinema"],
          "parameters": [
            {
              "$ref": "#/components/parameters/CinemaPath"
            }
          ],
          "requestBody": {
            "$ref": "#/components/requestBodies/cinemaBody"
          },
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/screen": {
        "get": {
          "description": "Get list screens from database",
          "tags": ["Screen"],
          "parameters": [
            {
              "name": "cinemaId",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer"
              },
              "description": "Cinema id of this screen"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/screen"
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Server error"
            }
          }
        },
        "post": {
          "description": "Add cinema into database",
          "tags": ["Screen"],
          "parameters": [],
          "requestBody": {
            "$ref": "#/components/requestBodies/screenBody"
          },
          "responses": {
            "200": {
              "description": "OK"
            },
            "400": {
              "description": "Bad request"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
      "/screen/{screenId}": {
        "get": {
          "description": "Get screen inforamtion by screen id",
          "tags": ["Screen"],
          "parameters": [
            {
              "$ref": "#/components/parameters/ScreenPath"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application.json": {
                  "schema": {
                    "$ref": "#/components/schemas/screen"
                  }
                }
              }
            }
          }
        },
        "put": {
          "description": "Update screen information",
          "tags": ["Screen"],
          "parameters": [
            {
              "$ref": "#/components/parameters/ScreenPath"
            }
          ],
          "requestBody": {
            "$ref": "#/components/requestBodies/screenBody"
          },
          "responses": {
            "200": {
              "description": "OK"
            },
            "404": {
              "description": "Not Found"
            },
            "500": {
              "description": "Server Error"
            }
          }
        }
      }
    }
  }